test: refactored tests into separate files to parallelize execution and reduce runtime
test(ci): setup of matrix sharding to deal with test timeout
