+++ packages/studio-web/tests/fixtures/ref/edited/readalong.eaf
