+++ packages/studio-web/tests/fixtures/ref/readalong.eaf
