--- packages/studio-web/tests/ref/www/index.html
@@ -1,25 +1,0 @@
-
-        <!DOCTYPE html>
-        <html lang="en">
-            <head>
-                <meta charset="UTF-8">
-                <title>Sentence Paragraph Page</title>
-                <meta name="application-name" content="read along">
-                <meta name="generator" content="@readalongs/studio-web 1.5.1">
-                <meta name="viewport" content="width=device-width, initial-scale=1" />
-                <!-- Import fonts. Material Icons are needed by the web component -->
-                <link href="https://fonts.googleapis.com/css?family=Lato%7CMaterial+Icons%7CMaterial+Icons+Outlined" rel="stylesheet">
-            </head>
-
-            <body>
-                <!-- Here is how you declare the Web Component. Supported languages: en, fr -->
-                <read-along href="assets/sentence-paragr-20241023184608.readalong" audio="assets/sentence-paragr-20241023184608.wav" theme="light" language="en" image-assets-folder="assets/">
-                    <span slot='read-along-header'>Sentence Paragraph Page</span>
-                    <span slot='read-along-subheader'>by me</span>
-                </read-along>
-            </body>
-
-            <!-- The last step needed is to import the package -->
-            <script type="module" src='https://unpkg.com/@readalongs/web-component@^1.5.1/dist/web-component/web-component.esm.js'></script>
-        </html>
-
