--- packages/studio-web/tests/ref/www/sentence-paragr-date.txt
@@ -1,7 +1,0 @@
-This is a test.
-Sentence.
-
-Paragraph.
-
-
-Page.
